chat_flow/requirements.txt
# for audio
numpy       #==2.0.1
sounddevice #==0.4.7
soundfile   #==0.12.1
#webrtcvad   # replacement for ambient and threshold calculation

# openai
openai      ==1.78.0  
openai-agents[voice]

# environment secrets (api keys)
python-dotenv #==1.0.1


# in wsl
# sudo apt update && sudo apt install -y portaudio19-dev
# potentially also libasound2-dev